reduce compiler warnings

+++ src/dyn_server.c
@@ -444,7 +444,7 @@ server_pool_init(struct server_pool *sp, struct conf_pool *cp, struct context *c
     sp->client_connections = (uint32_t)cp->client_connections;
 
     sp->server_retry_timeout_ms = cp->server_retry_timeout_ms;
-    sp->server_failure_limit = (uint32_t)cp->server_failure_limit;
+    sp->server_failure_limit = (uint8_t)cp->server_failure_limit;
     sp->auto_eject_hosts = cp->auto_eject_hosts ? 1 : 0;
     sp->preconnect = cp->preconnect ? 1 : 0;
